Create pong.py
The code for a simple game of pong for two players. Player one uses the A and Z keys to move (A to move up, Z to move down), while player two uses the up and down arrow keys to move. The players use the keys to move their paddles on their respective sides to try and score points by passing the ball past the other player's paddle.